seeded hotels and updated migration to float

+++ waiwai_backend/waiwai_backend_2/db/migrate/20190914031812_create_hotels.rb
@@ -2,8 +2,8 @@ class CreateHotels < ActiveRecord::Migration[5.2]
   def change
     create_table :hotels do |t|
       t.text :name
-      t.integer :rate
-      t.integer :stars
+      t.float :rate
+      t.float :stars
       t.text :room
       t.text :description
       t.text :image
